롯데카드 신규대상여부가 N 인 경우 발급정보 삭제 조건 추가 (#286)

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/service/lotte/LotteIssuanceService.java
@@ -79,7 +79,10 @@ public class LotteIssuanceService {
 			repoD1200.save(Lotte_D1200.builder().idxCorp(userCorp.idx()).build());
 			return "SUCCESS";
 		} else if ("N".equals(resultOfD1000.getBzNewYn())) {
-			commonCardService.deleteAllIssuanceInfo(userCorp.user());
+			Lotte_D1100 d1100 = repoD1100.getTopByIdxCorpOrderByIdxDesc(userCorp.idx());
+			if (d1100 == null && CardCompany.LOTTE.equals(userCorp.user().cardCompany())) {
+				commonCardService.deleteAllIssuanceInfo(userCorp.user());
+			}
 			return "FAIL";
 		} else {
 			String msg = "bzNewYn is not Y/N. resultOfD1000.getBzNewYn() = " + resultOfD1000.getBzNewYn();
